Really fix build (cf r8796). It appears only chaste-bob saw a problem...
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk/notforrelease_cell_based@8799 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ src/tissue/VertexBasedTissue.cpp
@@ -439,7 +439,7 @@ void VertexBasedTissue<DIM>::WriteResultsToFiles()
 
         if (TissueConfig::Instance()->GetOutputCellAncestors())
         {
-            double ancestor_index = (p_cell->GetAncestor() == UNSIGNED_UNSET) ? -1 : p_cell->GetAncestor();
+            double ancestor_index = (p_cell->GetAncestor() == UNSIGNED_UNSET) ? (-1.0) : (double)p_cell->GetAncestor();
             cell_ancestors.push_back(ancestor_index);
         }
         if (TissueConfig::Instance()->GetOutputCellProliferativeTypes())
